make authorization and authentication for user

+++ app/expense.py
@@ -1,4 +1,5 @@
 from flask import Blueprint, jsonify, request
+from flask_jwt_extended import jwt_required, current_user
 from marshmallow import ValidationError
 
 from app.db import Expense, db
@@ -8,6 +9,7 @@ bp = Blueprint("expense", __name__, url_prefix="/expenses")
 
 
 @bp.route("/", methods=["GET"])
+@jwt_required()
 def get_expenses():
     """
     Повертає список усіх витрат
@@ -16,19 +18,28 @@ def get_expenses():
         - витрати
     produces:
         - application/json
+    parameters:
+    - name: Authorization
+      in: header
+      description: JWT токен
+      required: true
     responses:
-          200:
+        200:
             description: Список витрат
             schema:
                 type: array
                 items:
                     $ref: '#/definitions/ExpenseOut'
+        401:
+            description: Немає доступу
+            schema:
+                $ref: '#/definitions/Unauthorized'
     """
-    expenses = Expense.query.all()
-    return jsonify(expenses_schema.dump(expenses)), 200
+    return jsonify(expenses_schema.dump(current_user.expenses)), 200
 
 
 @bp.route("/", methods=["POST"])
+@jwt_required()
 def add_expense():
     """
     Створює нову витрату
@@ -38,6 +49,10 @@ def add_expense():
     produces:
         - application/json
     parameters:
+    - name: Authorization
+      in: header
+      description: JWT токен
+      required: true
     - name: expense
       in: body
       description: Дані витрати
@@ -49,19 +64,28 @@ def add_expense():
             description: Створена витрата
             schema:
                 $ref: '#/definitions/ExpenseOut'
+        401:
+            description: Немає доступу
+            schema:
+                $ref: '#/definitions/Unauthorized'
+        422:
+            description: Помилка валідації
     """
     json_data = request.json
     try:
         data = expense_schema.load(json_data)
     except ValidationError as err:
         return err.messages, 422
-    new_expense = Expense(title=data["title"], amount=data["amount"])
+    new_expense = Expense(
+        title=data["title"], amount=data["amount"], user_id=current_user.id
+    )
     db.session.add(new_expense)
     db.session.commit()
     return jsonify(expense_schema.dump(new_expense)), 201
 
 
 @bp.route("/<int:id>", methods=["GET"])
+@jwt_required()
 def get_expense(id):
     """
     Повертає одну витрату за ідентифікатором
@@ -71,6 +95,10 @@ def get_expense(id):
     produces:
         - application/json
     parameters:
+    - name: Authorization
+      in: header
+      description: JWT токен
+      required: true
     - name: id
       in: path
       description: Ідентифікатор витрати
@@ -81,16 +109,23 @@ def get_expense(id):
             description: Знайдена витрата
             schema:
                 $ref: '#/definitions/ExpenseOut'
+        401:
+            description: Немає доступу
+            schema:
+                $ref: '#/definitions/Unauthorized'
         404:
             description: Не знайдено витрату за ідентифікатором
             schema:
                 $ref: '#/definitions/NotFound'
     """
     expense = db.get_or_404(Expense, id)
+    if expense.user_id != current_user.id:
+        return (jsonify(error="У вас немає доступу до цієї витрати")), 401
     return jsonify(expense_schema.dump(expense)), 200
 
 
 @bp.route("/<int:id>", methods=["PATCH"])
+@jwt_required()
 def update_expense(id):
     """
     Оновлює дані витрати за ідентифікатором
@@ -100,6 +135,10 @@ def update_expense(id):
     produces:
         - application/json
     parameters:
+    - name: Authorization
+      in: header
+      description: JWT токен
+      required: true
     - name: id
       in: path
       description: Ідентифікатор витрати
@@ -116,12 +155,22 @@ def update_expense(id):
             description: Оновлена витрата
             schema:
                 $ref: '#/definitions/ExpenseOut'
+        401:
+            description: Немає доступу
+            schema:
+                $ref: '#/definitions/Unauthorized'
         404:
             description: Не знайдено витрату за ідентифікатором
             schema:
                 $ref: '#/definitions/NotFound'
+        422:
+            description: Помилка валідації
     """
     expense = db.get_or_404(Expense, id)
+
+    if expense.user_id != current_user.id:
+        return (jsonify(error="У вас немає доступу до цієї витрати")), 401
+
     json_data = request.json
     try:
         data = expense_schema.load(json_data, partial=True)
@@ -134,6 +183,7 @@ def update_expense(id):
 
 
 @bp.route("/<int:id>", methods=["DELETE"])
+@jwt_required()
 def delete_expense(id):
     """
     Видаляє витрату за ідентифікатором
@@ -143,6 +193,10 @@ def delete_expense(id):
     produces:
         - application/json
     parameters:
+    - name: Authorization
+      in: header
+      description: JWT токен
+      required: true
     - name: id
       in: path
       description: Ідентифікатор витрати
@@ -151,12 +205,20 @@ def delete_expense(id):
     responses:
         204:
             description: Успішне видалення витрати
+        401:
+            description: Немає доступу
+            schema:
+                $ref: '#/definitions/Unauthorized'
         404:
             description: Не знайдено витрату за ідентифікатором
             schema:
                 $ref: '#/definitions/NotFound'
     """
     expense = db.get_or_404(Expense, id)
+
+    if expense.user_id != current_user.id:
+        return (jsonify(error="У вас немає доступу до цієї витрати")), 401
+
     db.session.delete(expense)
     db.session.commit()
     return "", 204
